backend/requirements.txt
# FastAPI and web framework
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.17

# Database
sqlalchemy==2.0.36

# AI and ML
google-genai==1.47.0

# Environment variables
python-dotenv==1.0.1

# Async file operations
aiofiles==25.1.0

# Browser automation
playwright==1.58.0

# WebSocket support
websockets==15.0.1

# HTTP client
httpx==0.28.1